Reorganize stellar changeset helpers
- Rename deps.go to chain_util.go and fold the byte converters
  (dataIDsToBytes, workflowNameToBytes, workflowOwnerToBytes) into it
- Move recordAddress next to its only callers in deploy.go
- Rename testutil_test.go to test_util.go
- Drop validation_test.go
