--- deployment/data-feeds/changeset/stellar/doc.go
@@ -1,3 +1,0 @@
-// Package stellar contains CLD changesets for the Stellar Data Feeds cache
-// and proxy contracts.
-package stellar
