+++ deployment/data-feeds/changeset/stellar/validation.go
@@ -1,11 +1,8 @@
 package stellar
 
 import (
-	"encoding/hex"
 	"errors"
 	"fmt"
-	"math/big"
-	"strings"
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/datastore"
 	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
@@ -45,47 +42,4 @@ func verifyFeedPreconditions(env cldf.Environment, chainSel uint64, qualifier, v
 	}
 	_, err := dataIDsToBytes(dataIDs)
 	return err
-}
-
-// dataIDsToBytes converts hex feed IDs to [16]byte. Data IDs are canonically
-// left-aligned, so short values are left-justified with trailing zero padding.
-func dataIDsToBytes(ids []string) ([][16]byte, error) {
-	out := make([][16]byte, 0, len(ids))
-	for _, id := range ids {
-		v, ok := new(big.Int).SetString(id, 0)
-		if !ok {
-			return nil, fmt.Errorf("invalid data_id: %q", id)
-		}
-		if v.BitLen() > 128 {
-			return nil, fmt.Errorf("data_id too long: %q (%d bits)", id, v.BitLen())
-		}
-		var b [16]byte
-		copy(b[:], v.Bytes())
-		out = append(out, b)
-	}
-	return out, nil
-}
-
-// workflowNameToBytes right-pads an ASCII workflow name into [10]byte.
-func workflowNameToBytes(s string) ([10]byte, error) {
-	var out [10]byte
-	if len(s) > len(out) {
-		return out, fmt.Errorf("workflow name %q exceeds %d bytes", s, len(out))
-	}
-	copy(out[:], s)
-	return out, nil
-}
-
-// workflowOwnerToBytes decodes a 20-byte hex workflow owner.
-func workflowOwnerToBytes(hexStr string) ([20]byte, error) {
-	var out [20]byte
-	b, err := hex.DecodeString(strings.TrimPrefix(hexStr, "0x"))
-	if err != nil {
-		return out, fmt.Errorf("invalid workflow owner %q: %w", hexStr, err)
-	}
-	if len(b) != len(out) {
-		return out, fmt.Errorf("workflow owner must be %d bytes, got %d", len(out), len(b))
-	}
-	copy(out[:], b)
-	return out, nil
 }
